Use a full-page landing for get-involved
